Updated logging to use new sdk function
